// src/shared/infraestructure/enums/message.enum.ts
export enum MessageEnum {
  CREDENTIALS_OK = 'Autenticación Correcta.!',
  CREDENTIALS_INVALID = 'Las credenciales son incorrectas!',
  ACCESS_INVALID = 'El acceso es inválido!',
  PASSWORD_UPDATED = 'Contraseña Actualizada!',
  PASSWORD_RESET = 'Contraseña reseteada!',  
  LOGIN_NOT_AUTORIZED = 'No autorizado!!!',  
  
  TOKEN_MODIFIED = "Token modificado",
  TOKEN_NOT_EXIST = "Token No Existe!!",
  TOKEN_DELETE = "Token eliminado correctamente.",

  ENTITY_SELECT = 'Datos devueltos correctamente.',
  ENTITY_ERROR_CREATE = 'Los datos no pudieron ser almacenados correctamente.',
  ENTITY_SELECT_EMPTY = 'No se encontraron registros.',
  ENTITY_PROCESS = 'Datos procesados correctamente.',

  USER_CREATED = 'Datos del Usuario fueron creados correctamente.',
  USER_UPDATE = 'Datos del Usuario fueron modificados correctamente.',  
  USER_DELETE = 'Datos del Usuario fueron eliminados correctamente.',    
  USER_EXIST = 'Datos de Usuario ya Existen!',
  USER_NOT_EXIST = 'Datos de Usuario no existen!',
  USER_VALIDATED = 'Datos del Usuario fueron validados correctamente.',
  
  PLAYER_CREATED = 'Datos del jugador fueron creados correctamente',
  PLAYER_NOT_CREATE = 'Datos del Jugador no fueron creados',
  PLAYER_EXIST = 'Datos de Jugador ya Existen!',
  PLAYER_NOT_EXIST = 'Datos de Jugador no Existen!',

  TEAM_EXIST = 'Datos del equipo ya Existen!',
  TEAM_NOT_EXIST = 'Datos del equipo no Existen!',
  TEAM_CREATED = 'Datos del equipo fueron creados correctamente',
  TEAM_UPDATE = 'Datos del equipo fueron modificados correctamente',
  TEAM_REGISTER = 'Datos del equipo fueron registrados correctamente',

  CHAMPIONSHIP_CREATED = 'Datos del campeonato fueron creados correctamente',
  CHAMPIONSHIP_UPDATE = 'Datos del campeonato fueron modificados correctamente',
  CHAMPIONSHIP_EXIST = 'Datos del campeonato ya Existen!',
  CHAMPIONSHIP_NOT_EXIST = 'Datos del campeonato no Existen!',
  CHAMPIONSHIP_ERROR_DATE = 'Error en las fechas',

  SCORE_CREATED = 'Datos del score fueron creados correctamente',
  SCORE_UPDATE = 'Datos del score fueron modificados correctamente',
  SCORE_EXIST = 'Datos del resultado del partido ya existen!',
  SCORE_NOT_EXIST = 'Datos del resultado del partido no existen!',

  PROGRAMMING_CREATED = 'Datos de la programación fueron creados correctamente',
  PROGRAMMING_UPDATE = 'Datos de la programación fueron modificados correctamente',
  PROGRAMMING_EXIST = 'Datos de la programación ya existen!',
  PROGRAMMING_NOT_EXIST = 'Datos de la programación no existen!',

  DISCIPLINES_NOT_EXIST = 'No se encontraron las disciplinas',

}